# Conflicts:
#	src/main/java/net/fabricmc/loom/configuration/ide/RunConfig.java
#	src/main/java/net/fabricmc/loom/configuration/ide/RunConfigSettings.java

+++ src/main/java/net/fabricmc/loom/configuration/providers/minecraft/mapped/AbstractMappedMinecraftProvider.java
@@ -173,15 +173,19 @@ public abstract class AbstractMappedMinecraftProvider<M extends MinecraftProvide
 	}
 
 	protected String getName(MinecraftJar.Type type) {
-		final String intermediateName = extension.getIntermediateMappingsProvider().getName();
-
 		var sj = new StringJoiner("-");
 		sj.add("minecraft");
 		sj.add(type.toString());
 
-		// Include the intermediate mapping name if it's not the default intermediary
-		if (!intermediateName.equals(IntermediaryMappingsProvider.NAME)) {
-			sj.add(intermediateName);
+		if (!extension.disableObfuscation()) {
+			// Include the intermediate mapping name if it's not the default intermediary
+			final String intermediateName = extension.getIntermediateMappingsProvider().getName();
+
+			if (!intermediateName.equals(IntermediaryMappingsProvider.NAME)) {
+				sj.add(intermediateName);
+			}
+		} else {
+			sj.add("deobf");
 		}
 
 		if (getTargetNamespace() != MappingsNamespace.NAMED) {
@@ -192,6 +196,10 @@ public abstract class AbstractMappedMinecraftProvider<M extends MinecraftProvide
 	}
 
 	protected String getVersion() {
+		if (extension.disableObfuscation()) {
+			return extension.getMinecraftProvider().minecraftVersion();
+		}
+
 		return "%s-%s".formatted(extension.getMinecraftProvider().minecraftVersion(), extension.getMappingConfiguration().mappingsIdentifier());
 	}
 
@@ -237,22 +245,29 @@ public abstract class AbstractMappedMinecraftProvider<M extends MinecraftProvide
 		}
 	}
 
-	private void remapJar(RemappedJars remappedJars, ConfigContext configContext) throws IOException {
+	protected void remapJar(RemappedJars remappedJars, ConfigContext configContext) throws IOException {
+		if (extension.disableObfuscation()) {
+			// TODO debof - can we skip this?
+			Files.createDirectories(remappedJars.outputJarPath().getParent());
+			Files.copy(remappedJars.inputJar(), remappedJars.outputJarPath(), StandardCopyOption.REPLACE_EXISTING);
+			getMavenHelper(remappedJars.type()).savePom();
+			return;
+		}
+
 		final MappingConfiguration mappingConfiguration = extension.getMappingConfiguration();
 		final String fromM = remappedJars.sourceNamespace().toString();
 		final String toM = getTargetNamespace().toString();
 
 		Files.deleteIfExists(remappedJars.outputJarPath());
 
+		final AnnotationsData remappedAnnotations = AnnotationsData.getRemappedAnnotations(getTargetNamespace(), mappingConfiguration, getProject(), configContext.serviceFactory(), toM);
 		final Map<String, String> remappedSignatures = SignatureFixerApplyVisitor.getRemappedSignatures(getTargetNamespace() == MappingsNamespace.INTERMEDIARY, mappingConfiguration, getProject(), configContext.serviceFactory(), toM);
 		final MinecraftVersionMeta.JavaVersion javaVersion = minecraftProvider.getVersionInfo().javaVersion();
 		final boolean fixRecords = javaVersion != null && javaVersion.majorVersion() >= 16;
 
 		TinyRemapper remapper = TinyRemapperHelper.getTinyRemapper(getProject(), configContext.serviceFactory(), fromM, toM, fixRecords, (builder) -> {
-			AnnotationsData annotationsData = mappingConfiguration.getAnnotationsData();
-
-			if (annotationsData != null) {
-				builder.extraPostApplyVisitor(new AnnotationsApplyVisitor(annotationsData));
+			if (remappedAnnotations != null) {
+				builder.extraPostApplyVisitor(new AnnotationsApplyVisitor(remappedAnnotations));
 			}
 
 			builder.extraPostApplyVisitor(new SignatureFixerApplyVisitor(remappedSignatures));
